% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\title{Estimate Area}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{A \code{numeric} (integer) used to denote the number of simulations}

\item{seed}{seed for reproducibility}
}
\value{
A\code{list} with estimate area and if the point is inside or outside the area
}
\description{
Calculate the area of the square
}
