% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot.area}
\alias{plot.area}
\title{Plot Area}
\usage{
\method{plot}{area}(x)
}
\value{
A\code{graph}
}
\description{
help visualize the plot
}
\examples{
plot.area(estimate_area(B= 5000))
}
\author{
Enzo Beijer
}
