% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot.area}
\alias{plot.area}
\title{Plot Area}
\usage{
\method{plot}{area}(x)
}
\arguments{
\item{x}{value returning from the estimation of the area}
}
\value{
A\code{graph}
}
\description{
help visualize the plot
}
